@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -29,24 +29,24 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.0
-  shared_preferences: ^0.5.12
+  cupertino_icons: ^1.0.4
+  shared_preferences: ^2.0.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  build_runner: ^1.4.0
-  flutter_slidable: 0.5.4
-  animated_text_kit: 2.4.0
-  bottom_navy_bar: ^5.5.0
-  font_awesome_flutter: 8.12.0
+  build_runner: ^2.0.0
+  flutter_slidable: ^0.6.0
+  animated_text_kit: ^4.0.0
+  bottom_navy_bar: ^6.0.0
+  font_awesome_flutter: ^9.0.0
   fair_compiler:
     path: ../../compiler
 
 # Switch to another stable flutter version
 dependency_overrides:
   fair_version:
-    path: ../../flutter_version/flutter_1_17_3
+    path: ../../flutter_version/flutter_2_5_0
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
 
